Added printing helper functions

+++ ft_putnbr.c
@@ -6,13 +6,13 @@
 /*   By: dslogrov <dslogrove@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/22 13:13:17 by dslogrov          #+#    #+#             */
-/*   Updated: 2018/05/25 14:02:31 by dslogrov         ###   ########.fr       */
+/*   Updated: 2019/06/27 16:43:37 by dslogrov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr(int n)
+void	ft_putnbr(intmax_t n)
 {
 	ft_putnbr_fd(n, 1);
 }
